Reorganize error handling
